+++ fstrack_tractor_app/test/features/home/presentation/pages/home_page_test.dart
@@ -4,6 +4,7 @@ import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:fstrack_tractor/core/network/connectivity_checker.dart';
 import 'package:fstrack_tractor/core/theme/app_theme.dart';
 import 'package:fstrack_tractor/features/auth/domain/entities/user_entity.dart';
 import 'package:fstrack_tractor/features/auth/presentation/bloc/auth_bloc.dart';
@@ -15,6 +16,7 @@ import 'package:fstrack_tractor/features/home/presentation/widgets/greeting_head
 import 'package:fstrack_tractor/features/home/presentation/widgets/role_based_menu_cards.dart';
 import 'package:fstrack_tractor/injection_container.dart';
 import 'package:mocktail/mocktail.dart';
+import '../../../../mocks/mock_connectivity_checker.dart';
 import 'package:fstrack_tractor/features/weather/presentation/bloc/weather_bloc.dart';
 import 'package:fstrack_tractor/features/weather/presentation/bloc/weather_event.dart';
 import 'package:fstrack_tractor/features/weather/presentation/bloc/weather_state.dart';
@@ -31,6 +33,7 @@ const Duration _weatherWidgetInitDelay = Duration(milliseconds: 100);
 void main() {
   late MockAuthBloc mockAuthBloc;
   late MockWeatherBloc mockWeatherBloc;
+  late MockConnectivityChecker mockConnectivityChecker;
   late StreamController<WeatherState> weatherStreamController;
 
   setUpAll(() {
@@ -43,6 +46,7 @@ void main() {
     mockAuthBloc = MockAuthBloc();
 
     mockWeatherBloc = MockWeatherBloc();
+    mockConnectivityChecker = MockConnectivityChecker();
     weatherStreamController = StreamController<WeatherState>.broadcast();
     when(() => mockWeatherBloc.state).thenReturn(const WeatherLoading());
     when(() => mockWeatherBloc.stream).thenAnswer((_) => weatherStreamController.stream);
@@ -55,16 +59,25 @@ void main() {
       getIt.unregister<WeatherBloc>();
     }
     getIt.registerSingleton<WeatherBloc>(mockWeatherBloc);
+
+    if (getIt.isRegistered<ConnectivityChecker>()) {
+      getIt.unregister<ConnectivityChecker>();
+    }
+    getIt.registerSingleton<ConnectivityChecker>(mockConnectivityChecker);
   });
 
   tearDown(() async {
     await weatherStreamController.close();
     await mockWeatherBloc.close();
+    mockConnectivityChecker.dispose();
   });
 
   tearDownAll(() {
     if (getIt.isRegistered<WeatherBloc>()) {
       getIt.unregister<WeatherBloc>();
+    }
+    if (getIt.isRegistered<ConnectivityChecker>()) {
+      getIt.unregister<ConnectivityChecker>();
     }
     getIt.allowReassignment = false;
   });